Playing Game
To play the game simply download the exe. file above to play the game. The user is given seven tries to guess what a word is. Each time a guess is incorrect a part of the hangman appears. If the user tries to guess the word all at once. A part of the hangman will appear for every wrong letter, but not for the correct ones.  The user will be able to see what the word looks like after every try.